requirements.txt
Django==2.0.3
mysqlclient==1.3.12
django-recaptcha
boto3==1.4.4
gunicorn
image
ujson
